+++ bankid-frontend/src/views/AutoStartView.vue
